// Servers/models/ISO-42001/annexCategoryStructISO.model.ts
import { Column, DataType, ForeignKey, Model, Table } from "sequelize-typescript";
import { AnnexStructISOModel } from "./annexStructISO.model";

export type AnnexCategoryStructISO = {
  id?: number; // auto generated by database
  title: string;
  description: string;
  guidance: string;
  sub_id: number;
  order_no?: number;
  annex_id?: number;
}

@Table({
  tableName: "annexcategory_struct_iso"
})
export class AnnexCategoryStructISOModel extends Model<AnnexCategoryStructISO> {
  @Column({
    type: DataType.INTEGER,
    autoIncrement: true,
    primaryKey: true,
  })
  id!: number;

  @Column({
    type: DataType.STRING
  })
  title!: string;

  @Column({
    type: DataType.STRING
  })
  description!: string;

  @Column({
    type: DataType.STRING
  })
  guidance!: string;

  @Column({
    type: DataType.INTEGER
  })
  sub_id!: number;

  @Column({
    type: DataType.INTEGER
  })
  order_no?: number;

  @ForeignKey(() => AnnexStructISOModel)
  @Column({
    type: DataType.INTEGER
  })
  annex_id?: number;
}